+++ packages/openapi/src/openapi-request-builder.ts
@@ -14,7 +14,8 @@ import {
 } from '@sap-cloud-sdk/http-client';
 import {
   filterCustomRequestConfig,
-  OriginOptions
+  OriginOptions,
+  encodeTypedClientRequest
 } from '@sap-cloud-sdk/http-client/internal';
 
 /**
@@ -103,6 +104,7 @@ export class OpenApiRequestBuilder<ResponseT = any> {
         url: this.getPath(),
         headers: this.getHeaders(),
         params: this.getParameters(),
+        parameterEncoder: encodeTypedClientRequest,
         data: this.parameters?.body
       },
       { fetchCsrfToken }